requirements.txt
six
kiwisolver>=1.0.1
python-dateutil>=2.1
cycler>=0.10
pyparsing!=2.0.4,!=2.1.2,!=2.1.6,>=2.0.3
pillow
matplotlib>=3.1.3
pyyaml
numpy